@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-ruby "2.0.0"
+ruby "2.1.2"
 gem 'rails', '4.1.6'
 
 # Use Bootstrap
